requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
astroid==1.6.3
billiard==3.3.0.23
celery==3.1.26.post2
cffi==1.11.5
colorama==0.3.9
cryptography==2.2.2
diff-match-patch==20121119
Django==2.0.4
django-appconf==1.0.2
django-celery==3.2.2
django-ckeditor==5.4.0
django-crispy-forms==1.7.2
django-formtools==2.1
django-imagekit==4.0.2
django-import-export==1.0.1
django-js-asset==1.1.0
django-reversion==3.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.7
isort==4.3.4
jdcal==1.4
kombu==3.0.37
lazy-object-proxy==1.3.1
lxml==4.2.1
mccabe==0.6.1
odfpy==1.3.6
openpyxl==2.5.5
pilkit==2.0
Pillow==5.1.0
pycparser==2.18
pylint==1.8.4
PyMySQL==0.9.2
pytz==2018.4
PyYAML==3.13
redis==2.10.6
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
